+++ TinyWebServer/src/tcpsocket.h
@@ -15,7 +15,7 @@ public:
     ~TcpSocket() override;
 
     void read(std::string& buffer) override;
-    int write(const char* buf, size_t size) override;
+    int write(const char* buf, int size) override;
 
     void close() override;
